Ticket #467: Optimize the data-type engine for heterogeneous cases
 * Submitted by: George Bosilca
 * Reviewed by: Jeff Squyres
 * Merged by: Tim Mattox

Merged r11945, r11962, r11972, r11986, which closes trac:467:

r11945

Fine grain selection for heterogeneous environments. The hetero
version of the conversion function are more complex and costly than a
simple memcpy. Therefore, we want to decrease as much as possible the
usage of these functions.  We now check not only th HOMOGENEOUS flag
on the datatype or convertor, but the bits indicating a type is in
use. If a communication transfert a type having the same
representation on both peers we can use the optimized version of the
conversion.  In same time we build a more accurate conversion table
for each master convertor, based on the minimum differences between
the 2 architectures.

r11962

With the lastest changes for heterogeneous environments we have to
compute the hetero flag a little bit different. The hetero flag is now
attached to a convertor if and only if we can use the optimized
conversion functions. It's a little bit broader than before (the 2
architectures has to be identical).

r11972

Mea culpa ... that was an weird error. When playing with 64 bits types
we have to the constant before using the << operator ... otherwise
everything is 32 bits ...  which does not give the right answer.

r11986

Cast to the uint64_t before playing with the << operator in order to
make sure we keep all usefull informations.

This commit was SVN r12028.

The following SVN revision numbers were found above:
  r11945 --> open-mpi/ompi@f4da7a80bd194f97fae7a23f0a161de519cfb4b2
  r11962 --> open-mpi/ompi@d05f4929018025f947f68222e97526c3c8c856a3
  r11972 --> open-mpi/ompi@e9572c7aee46b0fe3dd0c544e578bbec080b4c81
  r11986 --> open-mpi/ompi@bf1d4ee2598ef84fe53bcc97d1685a0d912483a8

The following Trac tickets were found above:
  Ticket 467 --> https://svn.open-mpi.org/trac/ompi/ticket/467

+++ ompi/datatype/dt_module.c
@@ -50,17 +50,17 @@ int ompi_ddt_dfd = -1;
 #define INIT_BASIC_TYPE( TYPE, NAME )                                        \
     { BASEOBJ_DATA, 0/*size*/, 0 /*align*/, 0/*true_lb*/, 0/*true_ub*/,      \
             0/*lb*/, 0/*ub*/, DT_FLAG_PREDEFINED, TYPE, 1,                   \
-      (((unsigned long long)1)<<(TYPE)), EMPTY_DATA(NAME) }
+      (((uint64_t)1)<<(TYPE)), EMPTY_DATA(NAME) }
 
 #define INIT_BASIC_DATA( TYPE, ALIGN, NAME, FLAGS )                                \
     { BASEOBJ_DATA, sizeof(TYPE), ALIGN, 0, sizeof(TYPE),                          \
             0, sizeof(TYPE), DT_FLAG_BASIC | (FLAGS),                              \
-            DT_##NAME, 1, (((unsigned long long)1)<<(DT_##NAME)), EMPTY_DATA(NAME) }
+            DT_##NAME, 1, (((uint64_t)1)<<(DT_##NAME)), EMPTY_DATA(NAME) }
 
 #define INIT_BASIC_DATA_WITH_NAME( TYPE, ALIGN, INTERNAL_NAME, NAME, FLAGS )       \
     { BASEOBJ_DATA, sizeof(TYPE), ALIGN, 0, sizeof(TYPE),                          \
             0, sizeof(TYPE), DT_FLAG_BASIC | (FLAGS),                              \
-            DT_##INTERNAL_NAME, 1, (((unsigned long long)1)<<(DT_##INTERNAL_NAME)),\
+            DT_##INTERNAL_NAME, 1, (((uint64_t)1)<<(DT_##INTERNAL_NAME)),          \
             EMPTY_DATA(NAME) }
 
 #define INIT_UNAVAILABLE_DATA( NAME )                                           \
@@ -76,7 +76,7 @@ int ompi_ddt_dfd = -1;
     { BASEOBJ_DATA, SIZE, ALIGN, 0/*true_lb*/, SIZE/*true_ub*/,                                    \
             0/*lb*/, SIZE/*ub*/,                                                                   \
             DT_FLAG_BASIC | DT_FLAG_DATA_FORTRAN | (FLAGS), \
-            (TYPE), 1, (((unsigned long long)1)<<(TYPE)), EMPTY_DATA(NAME) }
+            (TYPE), 1, (((uint64_t)1)<<(TYPE)), EMPTY_DATA(NAME) }
 #else
 #define INIT_BASIC_FORTRAN_TYPE( TYPE, NAME, SIZE, ALIGN, FLAGS )                                          \
     INIT_BASIC_TYPE( TYPE, NAME )
@@ -731,7 +731,7 @@ static inline int __dt_contain_basic_datatypes( const ompi_datatype_t* pData, ch
 {
     int i;
     size_t index = 0;
-    unsigned long long mask = 1;
+    uint64_t mask = 1;
 
     if( pData->flags & DT_FLAG_USER_LB ) index += snprintf( ptr, length - index, "lb " );
     if( pData->flags & DT_FLAG_USER_UB ) index += snprintf( ptr + index, length - index, "ub " );
